Move NavigatorCardBody to a new file
This makes `app.jsx` two times smaller then it was and we can now easier
port the navigatorCardBody to Typescript.
